package rename riscv -> riscv64 for consistency with GOARCH
